+++ R/data_documentation.R
@@ -1,0 +1,40 @@
+# cause of death ----
+#' Cause of death code
+#'
+#' This dataset contains the European shortlist for cause of death (English, Norwegian)
+#'
+#' @format
+#' \describe{
+#' \item{code}{Index}
+#' \item{description_en}{Description in English}
+#' \item{description_no}{Description in Norwegian}
+#' \item{code_icd10}{ICD-10 code range}
+#' }
+"cause_of_death"
+
+
+# icd10 Norwegian ----
+#' ICD-10 code in Norwegian
+#'
+#' This dataset contains the ICD-10 code in Norwegian.
+#'
+#' @format
+#' \describe{
+#' \item{code}{ICD-10 code}
+#' \item{description_no}{Description in Norwegian}
+#' }
+"icd10_no"
+
+
+# icd10 English ----
+#' ICD-10 code in English
+#'
+#' This dataset contains the ICD-10 code in English.
+#'
+#' @format
+#' \describe{
+#' \item{code}{ICD-10 code}
+#' \item{description_en}{Description in English}
+#' }
+"icd10_en"
+