+++ NEWS.md
@@ -1,3 +1,9 @@
-# medicode 2020.2.17
+# medicode 2023.02.27
+
+- Include `cause_of_death`, the European shortlist of cause of death
+- Include `icd10_no`, the ICD-10 code for Norwegian language
+- Include `icd10_en`, the ICD-10 code in English
+
+# medicode 2022.12.07
 
 - Package is created